oauth google and facebook login

+++ server.py
@@ -3,7 +3,18 @@ import json
 from google.oauth2 import service_account
 from google.cloud import translate
 from oauth2client.client import GoogleCredentials #for google cloud authentication
-import ndb #National Nutrient Database - United States Department of Agriculture
+
+# added due to google oauth login feature
+from oauth2client.client import OAuth2WebServerFlow #flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+from flask import make_response
+import requests
+import random, string
+from flask import session as login_session # a dictionary to store information
+
+# National Nutrient Database - United States Department of Agriculture
+import ndb
 
 from datetime import datetime  #required for method "now" for file name change
 from flask import Flask, render_template, request, url_for, redirect, flash, jsonify, send_from_directory
@@ -11,6 +22,7 @@ from werkzeug.utils import secure_filename  #required for file upload
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from food_database import (Base,
+                   User,
                    UOM,
                    Meal,
                    Ingredient,
@@ -48,14 +60,20 @@ jsonString = """{
   "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
   "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/serviceaccount-owner%40long-memory-188919.iam.gserviceaccount.com"
 }"""
-# d = { dictionary ...}
-#jsonString = json.dumps(d)
 service_account_info = json.loads(jsonString)
 credentials = service_account.Credentials.from_service_account_info(
     service_account_info)
 
 # Google translate client
 translateClient = translate.Client(target_language='en',credentials=credentials)
+
+# Google oauth credentials
+GOOGLE_WEB_CLIENT_ID = os.environ.get('GOOGLE_WEB_CLIENT_ID')
+GOOGLE_CLIENT_SECRET = os.environ.get('GOOGLE_CLIENT_SECRET')
+
+# Facebook oauth credentials
+FACEBOOK_APP_ID = os.environ.get('FACEBOOK_APP_ID')
+FACEBOOK_SECRET_KEY = os.environ.get('FACEBOOK_SECRET_KEY')
 
 # NDB connection
 NDB_KEY = os.environ.get('NDB_KEY')
@@ -98,6 +116,240 @@ def upload_static(filename):
     return send_from_directory('upload', filename)
 
 # Application Routes
+
+@app.route('/login')
+def showLogin():
+    # Generate a random session id and pass it to the login template
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits) for x in range(32))
+    login_session['state'] = state
+    return render_template("login.html", STATE=state, CLIENT_ID=GOOGLE_WEB_CLIENT_ID)
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        # TODO: Replace json file with path variable?
+        #oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        #oauth_flow.redirect_uri = 'postmessage'
+        oauth_flow = OAuth2WebServerFlow(client_id=GOOGLE_WEB_CLIENT_ID,
+                           client_secret=GOOGLE_CLIENT_SECRET,
+                           scope='',
+                           redirect_uri='postmessage')
+
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        print("Failed to upgrade the authorization code")
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != GOOGLE_WEB_CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print("Token's client ID does not match app's.")
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already connected.'),
+                                 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+    login_session['provider'] = 'google'
+
+    # Validates if user already exists in the database
+    # if not it creates a new account in the Database
+    if not getUserID(login_session['email']):
+        createUser(login_session)
+
+    login_session['user_id'] = getUserID(login_session['email'])
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print("done!")
+    return output
+
+
+# DISCONNECT - Revoke a current user's token and reset their login_session
+@app.route('/gdisconnect')
+def gdisconnect():
+    # Only disconnect a connected user.
+    access_token = login_session.get('access_token')
+    if access_token is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    if result['status'] == '200':
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+            # TODO: Logout / disconnect does not work correctly unless the browser is closed / the cookie is no longer valid
+            # at the moment it will show as failed to revoke token with a status 400 from google
+    else:
+        response = make_response(json.dumps('Failed to revoke token for given user.'), 400)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data.decode("utf-8") # .decode("utf-8") added by soeren
+
+    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
+        FACEBOOK_APP_ID, FACEBOOK_SECRET_KEY, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1].decode("utf-8") # .decode("utf-8") added by soeren
+
+
+    # Use token to get user info from API
+    userinfo_url = "https://graph.facebook.com/v2.8/me"
+    '''
+        Due to the formatting for the result from the server token exchange we have to
+        split the token first on commas and select the first index which gives us the key : value
+        for the server access token then we split it on colons to pull out the actual token value
+        and replace the remaining quotes with nothing so that it can be used directly in the graph
+        api calls
+    '''
+    token = result.split(',')[0].split(':')[1].replace('"', '')
+
+    url = 'https://graph.facebook.com/v2.8/me?access_token=%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    # print "url sent for API access:%s"% url
+    # print "API JSON result: %s" % result
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
+
+    # The token must be stored in the login_session in order to properly logout
+    login_session['access_token'] = token
+
+    # Get user picture
+    url = 'https://graph.facebook.com/v2.8/me/picture?access_token=%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+
+    login_session['picture'] = data["data"]["url"]
+
+    # see if user exists
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+
+    flash("Now logged in as %s" % login_session['username'])
+    return output
+
+
+@app.route('/fbdisconnect')
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    # The access token must me included to successfully logout
+    access_token = login_session['access_token']
+    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'DELETE')[1]
+    return "you have been logged out"
+
+
+# Disconnect based on provider
+@app.route('/disconnect')
+def disconnect():
+    if 'provider' in login_session:
+        if login_session['provider'] == 'google':
+            gdisconnect()
+            del login_session['gplus_id']
+            del login_session['access_token']
+        if login_session['provider'] == 'facebook':
+            fbdisconnect()
+            del login_session['facebook_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        del login_session['provider']
+        print("You have successfully been logged out.")
+        return redirect('/')
+    else:
+        print("You were not logged in")
+        return redirect('/')
+
+
 @app.route('/')
 def showHome():
     return render_template("index.html")
@@ -105,8 +357,9 @@ def showHome():
 
 @app.route('/meals')
 def showMeals():
-    meals = session.query(Meal)
-    return render_template("meals.html",meals=meals,getIngredients=getIngredients)
+    meals = session.query(Meal).outerjoin(Meal.user)
+    #user = getUserInfo(login_session['user_id'])
+    return render_template("meals.html",meals=meals,getIngredients=getIngredients,loginSession=login_session)
 
 @app.route('/food_facts')
 def showFoodFacts():
@@ -115,6 +368,15 @@ def showFoodFacts():
 
 @app.route('/meals/add', methods=['GET','POST'])
 def addMeals():
+
+    #del login_session['username']
+
+    # required if no user has yet loged in so the dictionary does not have the key
+    if 'username' not in login_session:
+        return redirect ('/login')
+        # TODO: return to the intendet page after login or redirect to original URL using request.referrer
+
+
     if request.method == 'POST' and request.form['button'] == "Save":
 
         # ---- FILE HANDLING -----
@@ -147,7 +409,8 @@ def addMeals():
                        description=request.form['description'],
                        portions=request.form['portions'],
                        calories="410 - 780",
-                       image=filename)
+                       image=filename,
+                       user_id=login_session['user_id'])
         session.add(newMeal)
         session.commit()
 
@@ -170,7 +433,7 @@ def addMeals():
             print("detected source language = " + translation['detectedSourceLanguage'])
             foodID = getFood(translation['translatedText'])
             if not foodID:
-                foodID = ""
+                foodID = None
             newIngredient = Ingredient(quantity=request.form['quantity'+row],
                                        uom_id=request.form['uom'+row],
                                        meal_id=newMeal.id,
@@ -198,6 +461,11 @@ def addMeals():
 
 @app.route('/meals/delete/<int:meal_id>', methods=['GET','POST'])
 def deleteMeal(meal_id):
+
+    if 'username' not in login_session:
+        return redirect ('/login')
+        # TODO: return to the intendet page after login or redirect to original URL using request.referrer
+
     o = session.query(Meal).filter_by(id=meal_id).one()
     if request.method == 'POST' and request.form['button'] == "Delete":
         session.delete(o)
@@ -222,6 +490,31 @@ def showMeal(meal_id):
 def jsonMeals():
     items = session.query(Meal).all()
     return jsonify(Meals=[m.serialize for m in items])
+
+# ---------------- User Helper Functions ---------------------------
+
+# Creates a new user in the database
+def createUser(login_session):
+    newUser = User(name=login_session['username'], email=login_session[
+                   'email'], picture=login_session['picture'])
+    session.add(newUser)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+# Retrieves the user object
+def getUserInfo(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+# Retrieves the user id based on an email
+def getUserID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except:
+        return None
+
 
 
 def getFood(keyword):
@@ -292,8 +585,10 @@ def getFood(keyword):
             return food.id
         else:
             print("No results in NDB found")
+            return None
     else:
         print("Missing API Key in Environment Variable")
+        return None
 
 
 def getIngredients(meal_id):
